Add Minitest::DebugReporter plugin
So, this was kind of a waste of effort...

I was looking at the source of `Minitest` that was included with my
version of ruby, and didn't realize there was a `#prerecord` method
added to `Minitest::AbstractReporter` in newer versions:

https://github.com/seattlerb/minitest/commit/c797bafa

So most of this was coded without knowing about that...

Anyway, this basically does the same thing as the functionality added
there, and allows either to work regardless of what version you are
using.  Kinda hacky, but it should help narrow down the issues we are
having with travis locking up.